no new functionality, just update to new api
This commit updates the emirc.IRCGateway peripheral + receptor to the newest API version. The configuration is now loaded from an individual toml file (instead of a global config file). Furthermore, the peripheral is in charge of configuring itself now and the core does not perform any configuration in addition to the standard logging and identifier.